feat: Optimize MissionEscrow.initialize types (#46)
- Change `MissionEscrow.initialize` arguments to `uint96` (missionId, rewardAmount) and `uint64` (expiresAt) to match storage layout.
- Remove redundant internal checks in `MissionEscrow.initialize` as types enforce limits.
- Update `IMissionEscrow` interface.
- Update `MissionFactory` to cast arguments safely after validation.
- Fix broken `test/ReputationAttestations.t.sol` by adding missing mocks.

This optimization reduces redundant checks and improves type safety in the interface.
Gas impact: Slight increase (~12 gas) in benchmarks due to factory casting overhead vs check removal, but code is cleaner and types are more correct.

+++ src/interfaces/IMissionEscrow.sol
@@ -73,10 +73,10 @@ interface IMissionEscrow {
     // =============================================================================
 
     function initialize(
-        uint256 missionId,
+        uint96 missionId,
         address poster,
-        uint256 rewardAmount,
-        uint256 expiresAt,
+        uint96 rewardAmount,
+        uint64 expiresAt,
         address guild,
         bytes32 metadataHash,
         bytes32 locationHash,
